Reuse memory, code cleanup, loop unrolling
For decoding, the decoded string is always guaranteed to be shorter than the encoded string, so we can just put the decoded string in the encoded string's memory and throw in an early null terminator to save an allocation. This brings us down from about 2.5s benchmark time to 2.3s

General code cleanup, including using pointer arithmetic instead of indexing and removing unnecessary final block passes. Brings us down from 2.3s to 2.2s

Micromanaging Clang's loop unrolling in the hot loops gets us down from 2.2s to 2.0s

From 2.5s to 2.0s leaves us with a total performance improvement of 20% for this commit (on Clang at least, not on GCC)